// backend/src/libs/whatsAppOficial/IWhatsAppOficial.interfaces.ts
export interface IMetaMessageinteractive {
    type: 'button' | 'list';
    header?: IMetaMessageinteractiveHeader;
    body?: IMetaMessageinteractiveBody;
    footer?: IMetaMessageinteractiveFooter;
    action: IMetaMessageinteractiveAction;
}

export interface IMetaMessageinteractiveAction {
    sections?: Array<IMetaMessageinteractiveActionSections>;
    buttons?: Array<IMetaMessageinteractiveActionButtons>;
}

export interface IMetaMessageinteractiveActionButtons {
    type: 'reply';
    reply: IMetaMessageinteractiveActionButtonsReply;
}

export interface IMetaMessageinteractiveActionButtonsReply {
    id: string;
    title: string;
}

export interface IMetaMessageinteractiveActionSections {
    title: string;
    rows: IMetaMessageinteractiveActionSectionsRows[];
}

export interface IMetaMessageinteractiveActionSectionsRows {
    id: string;
    title: string;
    description?: string;
}

export interface IMetaMessageinteractiveHeader {
    type: 'text' | 'image';
    image?: IMetaMessageinteractiveHeaderImage;
    text?: string;
}

export interface IMetaMessageinteractiveHeaderImage {
    id: string;
}

export interface IMetaMessageinteractiveBody {
    text: string;
}

export interface IMetaMessageinteractiveFooter {
    text: string;
}